Day 2: Test a report in a single pass
Rather than iterating through the report 3 times (once to calculate
diffs, once to check for magnitude, and once to check for direction),
just iterate through it once.

Note that for part 2, we're still iterating a report multiple times to
test with each level removed, but each of _those_ tests now only
iterates once.