Day 6: Optimize part 2
Rather than follow the whole path from the guard's starting position for
each obstacle, we can start form the step just before the obstacle and
see if that loops. We already know the guard will get to that point
eventually on her path, because we're only choosing points on her
original path.

We do have to make sure that we only consider the first time she would
encounter that position on her path. If she loops back to a position on
the original path, if we'd added an obstacle in the way she wouldn't
necessarily do that same loop.