Day 9: Speed up part 1
Rather than looking through the whole list each time for the file block
to move, maintain a pointer to the last file we moved and continue the
search from there. Combined with stopping when the two pointers meet,
this massively speeds it up (from tens of seconds to tens of
milliseconds).