Day 12: Optimize part 2 by only walking the region
Instead of walking the bounding box for each region twice (once going
horizontally and once vertically), instead walk just the cells in the
region. We can just check if each cell has an edge in each direction
where the previous cell in that row/column doesn't have an edge. This
still gives us all of the edge->non edge transitions, but without so
much overhead.